+++ lib/app_screens/main/Home.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_app/app_screens/discover/Discover.dart';
 import 'package:flutter_app/app_screens/main/Body.dart';
-import 'package:flutter_app/app_screens/main/main_attributes/HomeAttributes.dart';
+import 'package:flutter_app/app_screens/main/main_attributes/HomeAppBar.dart';
 import 'package:flutter_app/app_screens/settings/SideMenu.dart';
 import 'package:flutter_app/shared/helpers/icomoon.dart';
 // ignore: implementation_imports
