+++ lib/shared/style.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+const double defaultPadding = 20;
+
+const Color textAccentColor = Color(0xff6a7f9e);
+
+const fontStyleContent = TextStyle(
+  decoration: TextDecoration.none, // that by default so it's useless too
+  fontSize: 17,
+  fontWeight: FontWeight.w700,
+);
