+++ lib/widgets/expanded/widgets_attribute/Details/widget_details.dart
@@ -14,15 +14,15 @@ class DetailsText extends StatelessWidget{
         children: <Widget>[
           MainFont(),
           transparent_divider(),
-          SubFont1(),
+          PlatForm(),
           transparent_divider(),
-          SubFont2(),
+          Progressions(),
           transparent_divider(),
-          SubFont3(),
+          Editor(),
           transparent_divider(),
-          SubFont4(),
+          ReleaseDate(),
           transparent_divider(),
-          SubFont5(),
+          Description(),
           transparent_divider(),
           SubFont6(),
           transparent_divider(),
