+++ lib/widgets/expanded/widgets_attribute/Settings/FontAttribute/Contents.dart
@@ -4,18 +4,12 @@ import 'package:flutter_app/widgets/expanded/widgets_attribute/Settings/FontAttr
 class LocalID extends StatelessWidget{
   @override
   Widget build(BuildContext context) {
-    return Align(
-      child: Row(
-        children: <Widget>[
-          RichText(
-            text: TextSpan(
-              style: FontStyleContent(),
-              children: <TextSpan>[
-                TextSpan(text: "Local ID: "),
-                TextSpan(text: "8VxqWO9pRBTvpLGxFXquloo97X13", style: FontStyleSubContent()),
-              ],
-            ),
-          ),
+    return RichText(
+      text: TextSpan(
+        style: FontStyleContent(),
+        children: <TextSpan>[
+          TextSpan(text: "Local ID: "),
+          TextSpan(text: "8VxqWO9pRBTvpLGxFXquloo97X13", style: FontStyleSubContent()),
         ],
       ),
     );
